requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
discord-interactions==0.3.0
discord-py-slash-command==3.0.1
discord.py==1.7.3
idna==3.3
multidict==5.2.0
pycparser==2.21
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
spotipy==2.19.0
typing-extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2
